Add files via upload
This is a python file of this project and you need to install necessary libraries to go ahead with the coding part. All the necessary libraries are mentioned at the top of the file so install it before you proceed ahead. Feel free to raise an issue if you find any error in the code or in the output which you are getting. 

*Disclaimer*
The image url or the image address should be in jpg or png format.