Add files via upload
The removed_background.jpg file is the original image in which we have removed the background image from the main object. In the changed_background.jpg we have replaced the background with a new background image of our own choice. The user can change the background image of their own choice and can pixelate the main object in the image more.